+++ src/hooks/useCaptchaSiteKey.ts
@@ -1,9 +1,0 @@
-import * as React from 'react'
-
-export const CaptchaSiteKeyContext = React.createContext<string | undefined>(
-  undefined,
-)
-
-export default function useCaptchaSiteKey() {
-  return React.useContext(CaptchaSiteKeyContext)
-}