+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- support for invisible ReCAPTCHA
+
 ## [6.6.0] - 2024-09-25
 
 ### Added